Feature/handler for adding images to products

+++ models/product.go
@@ -3,8 +3,8 @@ package models
 type Product struct {
 	Id            string `json:"id"`
 	Name          string `json:"name"`
-	UnitPrice     int    `json:"unitPrice"`
-	BulkPrice     int    `json:"bulkPrice"`
+	UnitPrice     int64  `json:"unitPrice"`
+	BulkPrice     int64  `json:"bulkPrice"`
 	Inventory     int    `json:"inventory"`
 	ItemsInPacket int    `json:"itemsInPacket"`
 	StoreId       string    `json:"storeId"`
